python_files/pyproject.toml
[project]
name = "agent-tools"
version = "0.1.0"
description = "AI Agent Orchestration Tools - Fork, swarm, and orchestrate Gemini agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
# Main CLI commands - available globally after install
orchestrator = "tools.orchestrator:main"
agent-fork = "tools.agent_fork:main"
context-summarizer = "tools.context_summarizer:main"
skill-manager = "skill.context_aware_skill_system:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "skill"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
